move layouts to components and config api to services
