+++ client/src/pages/admin/auth/register.js
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { useRouter } from "next/router";
 import Swal from "sweetalert2";
-import api from "@/utils/backendApi";
+import api from "@/services/backendApi";
 
 function RegisterPage() {
 	const router = useRouter();
