+++ client/src/pages/admin/categories/[id].js
@@ -1,8 +1,8 @@
-import AdminLayout from "@/layouts/AdminLayout";
+import AdminLayout from "@/components/layouts/AdminLayout";
 import { formatDate } from "@/utils/common";
 import { useRouter } from "next/router";
 import { useState, useEffect } from "react";
-import api from "@/utils/backendApi";
+import api from "@/services/backendApi";
 import Swal from "sweetalert2";
 
 function CategoryPage() {
