+++ client/src/pages/admin/categories/create.js
@@ -1,6 +1,6 @@
 import { useState, useRef } from "react";
-import AdminLayout from "@/layouts/AdminLayout";
-import api from "@/utils/backendApi";
+import AdminLayout from "@/components/layouts/AdminLayout";
+import api from "@/services/backendApi";
 import Swal from "sweetalert2";
 
 function CreateCategoryPage() {
