+++ client/src/pages/admin/categories/edit/[id].js
@@ -1,7 +1,7 @@
 import { useEffect, useState, useRef } from "react";
 import { useRouter } from "next/router";
-import AdminLayout from "@/layouts/AdminLayout";
-import api from "@/utils/backendApi";
+import AdminLayout from "@/components/layouts/AdminLayout";
+import api from "@/services/backendApi";
 import Swal from "sweetalert2";
 
 function EditCategoryPage() {
