+++ client/src/pages/admin/index.js
@@ -1,4 +1,4 @@
-import AdminLayout from "@/layouts/AdminLayout";
+import AdminLayout from "@/components/layouts/AdminLayout";
 
 function AdminPage() {
 	return (
