+++ client/src/pages/admin/posts/[id].js
@@ -1,8 +1,8 @@
-import AdminLayout from "@/layouts/AdminLayout";
+import AdminLayout from "@/components/layouts/AdminLayout";
 import { formatDate } from "@/utils/common";
 import { useRouter } from "next/router";
 import { useState, useEffect } from "react";
-import api from "@/utils/backendApi";
+import api from "@/services/backendApi";
 import Swal from "sweetalert2";
 
 function PostPage() {
@@ -76,7 +76,10 @@ function PostPage() {
 						</h1>
 						<div className="my-8">
 							<span className="text-lg font-bold mr-3">
-								{formatDate(post.createdAt, post.readingMinutes)}
+								{formatDate(
+									post.createdAt,
+									post.readingMinutes
+								)}
 							</span>
 						</div>
 						{post.imageUrl && (
