+++ client/src/pages/admin/posts/create.js
@@ -1,7 +1,7 @@
 import { useState, useRef, useEffect } from "react";
 import { WithContext as ReactTags } from "react-tag-input";
-import AdminLayout from "@/layouts/AdminLayout";
-import api from "@/utils/backendApi";
+import AdminLayout from "@/components/layouts/AdminLayout";
+import api from "@/services/backendApi";
 import dynamic from "next/dynamic"; // Import the dynamic function from Next.js
 import Swal from "sweetalert2";
