+++ client/src/pages/client/posts/index.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
-import api from "@/utils/backendApi";
-import ClientLayout from "@/layouts/ClientLayout";
+import api from "@/services/backendApi";
+import ClientLayout from "@/components/layouts/ClientLayout";
 import PostList from "@/components/PostListClient";
 
 const Index = () => {
